requirements.txt
captcha>=0.7.1
tqdm>=4.65.0
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
google
